refactor: split persistence/tests.rs into logical submodules
Refactor the monolithic persistence test file (1,211 lines) into smaller,
focused modules for improved maintainability and navigation.

Changes:
- Split tests.rs into tests/ directory with 3 files:
  - tests/sql_dump.rs: 12 tests for SQL dump functionality (473 lines)
  - tests/json_persistence.rs: 12 tests for JSON persistence (696 lines)
  - tests/mod.rs: Module declarations and documentation (14 lines)

Benefits:
- Improved maintainability: Easier to locate and modify specific test categories
- Better organization: Logical grouping of related tests
- Faster reviews: Smaller, focused files are easier to review
- Reduced merge conflicts: Changes to different test categories won't conflict

All 24 existing tests preserved with no changes to test logic or coverage.
Test results: 22 passed, 2 failed (pre-existing failures in main branch)

Closes #1355
